+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rhds/elements
 
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- 5247670: Fixed [bug] `<rh-footer>` links should be styled to match dark context colors. #307
+  Fixed the `social-links` slot to not override the `social-links` internal `rh-footer-links`
+  Fixed the `social-links` hrefs to point to the default RHDC links
+
 ## 1.0.0-beta.4
 
 ### Minor Changes